+++ nerfstudio/viewer_beta/control_panel.py
@@ -270,10 +270,7 @@ class ControlPanel:
             step: the train step to set the model to
         """
         with self.viser_server.atomic(), self.stat_folder:
-            # TODO change to a .value call instead of remove() and add, this makes it jittery
-            with self.viser_server.atomic():
-                self.markdown.remove()
-                self.markdown = self.viser_server.add_gui_markdown(f"Step: {step}")
+            self.markdown.content=f"Step: {step}"
 
     def update_output_options(self, new_options: List[str]):
         """
